comento codigo

+++ src/main/scala/funciones.scala
@@ -14,6 +14,10 @@ trait Modulo {
 
   def startFunction : Unit
 
+  /**
+   * Funcion = getFilesFrom
+   * Obtiene todos los archivos de un directorio dado.
+   */
   def getFilesFrom(dir: String): List[File] = {
     val d = new File(dir)
     if (d.exists && d.isDirectory){
@@ -24,6 +28,11 @@ trait Modulo {
     else List[File]()
   }
 
+
+  /**
+   * Funcion = deleteFolder
+   * Elimina los archivos recursivos a una carpeta dada.
+   */
   def deleteFolder(destino: File) : Boolean = {
     def deleteAll (file: File) : Boolean = {
       if (file.isDirectory)
@@ -45,32 +54,33 @@ class Ordenador (ruta : String, out : TextArea, dest : Option[String] = None) ex
   val folders : List[Folder] = ConfigFile.getFolderList
   val carpeta_destino : String = dest.getOrElse(ConfigFile.getFolder())
 
+
+  /**
+   * Funcion = startFunction
+   * Funcion principal que recupera los valores configurados y procesa cada archivo.
+   */
   def startFunction = {
-    // Cada vez que ejecuta esta funcionalidad, recupero la lista de filtros si es que lo modifico en tiempo de ejecucion
     val filtros : List[FileKeyword] = ConfigFile.getFilters
     val tamFiltros = filtros.size
-    // La secuencia empieza siempre en 000
     var i : Int = 0
-    // Meto los archivos de las carpetas en una lista
     restantes = getFiles(ruta).sorted
     total = restantes.size
-    // Recorro la lista de filtros de arriba hacia abajo
+
     for (i <- 0 to tamFiltros) {
       if (i == tamFiltros)
         // llego al ultimo, ejecuto los restantes
         filtrar(".", "", "INICIO")
       else
-        // Paso los valores de cada filtro de la lista, para que el metodo filtro filtre con esos valores
         filtrar(filtros(i).keyword, filtros(i).folder, filtros(i).match_pos)
     }
 
     val dest_folder = new File(ruta + "\\" + carpeta_destino)
 
     Try(deleteFolder(dest_folder)) match {
-      case Success(s) if (!s) => throw ErrorCrearCarpeta("No se pudo limpiar la carpeta destino")
       case Failure(e) => out.append(e.getMessage)
-      case Success(_) =>
-        val p = createFiles(dest_folder)
+      case Success(s) =>
+        if (!s) throw ErrorCrearCarpeta("No se pudo limpiar la carpeta destino")
+        val p : Int = createFiles(dest_folder)
         if (p > 0) {
           out.append(s"($p/$total) archivos fueron organizados\n")
           if (p == total)
@@ -80,49 +90,69 @@ class Ordenador (ruta : String, out : TextArea, dest : Option[String] = None) ex
             throw ErrorTerminarProceso("No se procesaron todos los archivos")
           }
         }
-
     }
   }
 
 
+  /**
+   * Funcion = getFiles
+   * Recupera todos los archivos de todas las carpetas involucradas para ordenarlos
+   * Devuelve una lista
+   */
   private def getFiles(r: String): List[File] = ConfigFile.getFolderList.flatMap(folder => getFilesFrom(r + "\\" + folder.name))
 
 
+
+  /**
+   * Funcion = filtrar
+   * En base a una palabra, se fija en todos los archivos que estan en la lista 'restantes'.
+   * Si encuentra algun archivo que matche con la palabra, se agrega a otra lista 'ordenados'.
+   * Al final, la lista 'restantes' se reduce con los que encontro y la lista 'ordenados' crece.
+   */
 	private def filtrar (word: String, folder: String, pos: String): Unit = {
-    // Funcion que veriica la posicion, el patron y el nombre de la carpeta origen con el destino
+    // Cada vez que encuentre un *** dentro del keyword, lo reemplazo con .*
+    val regex = word.replaceAllLiterally("***", ".*") + ".*"
+    val patron = Pattern.compile(regex)
+
+    /**
+     * Funcion = criteria
+     * Valida si el nombre de un archivo cumple con un regex.
+     */
     def criteria(x: File) : Boolean = {
-      // En caso que haya recorrido todos los filtros, los demas los dejo cumplir la condicion.
       if (folder.isEmpty) return true
-      // Cada vez que encuentre un *** dentro del keyword, lo reemplazo con .*
-      val regex = word.replaceAllLiterally("***", ".*") + ".*"
-      val patron = Pattern.compile(regex)
       val matcher = patron.matcher(x.getName)
       if (matcher.find) {
         val p = if (matcher.start == 0) "INICIO" else "MEDIO"
         x.getParentFile.getName == folder && p == pos
       } else false
     }
-    // La lista de los archivos filtrados para luego sacar una diferencia
+
     var filtrados: List[File] = Nil
     for (f <- restantes.filter(criteria).sortBy(_.getName)){
-      // Inserto en la lista el nuevo nombre del archivo
       ordenados = (f, getNewName(index, f.getParentFile.getName, f.getName)) ::  ordenados
       index += 1
       filtrados = f :: filtrados
     }
-    // Hago la diferencia de los restantes menos los filtrados
+
     restantes = restantes filterNot filtrados.contains
 	}
 
+
+  /**
+   * Funcion = createFiles
+   * Crea la carpeta y los archivos ordenados dentro con su secuencia asignada.
+   * Devuelve la cantidad procesada.
+   */
   private def createFiles(dir: File) : Int = {
     var procesados = 0
+    if (ordenados.size == 0) throw ErrorCrearCarpeta("No existen archivos para ordenar")
     if (dir.mkdir)
-      out.append(s"Directorio ${dir.getName} creado.\n")
+      out.append(s"Directorio '${dir.getName}' creado.\n")
     else
       throw ErrorCrearCarpeta("No se limpio la carpeta destino")
 
     out.append(s"Archivos creados:\n")
-    // CREO LOS ARCHIVOS CON SUS RESPECTIVOS NOMBRES
+
     for ((src, name) <- ordenados.sortBy(_._2)){
       val destin = new File(dir.getPath + "\\" + name)
       val fos = new FileOutputStream(destin)
@@ -138,15 +168,22 @@ class Ordenador (ruta : String, out : TextArea, dest : Option[String] = None) ex
     procesados
   }
 
-	private def getNewName (i: Int, prefix: String, filename: String) = {
-    // Verifico si debo mostrar el nombre de carpeta de origen de archivo, contenido en su nombre.
+  /**
+   * Funcion = getNewName
+   * Construye el nuevo nombre del archivo, concatenando el indice con el prefijo (si es necesario) y su nombre original
+   */
+	private def getNewName (i: Int, prefix: String, filename: String) : String = {
     if (folders.filter(_.name == prefix).map(_.show).head)
       getIndex(i,total) + "_" + prefix + "_" + filename
     else
       getIndex(i,total) + "_" + filename
   }
 
-  // Creo el indice de los archivos
+
+  /**
+   * Funcion = getIndex
+   * Construye el indice del archivo
+   */
 	private def getIndex(i: Int, max: Int) = "0" * (max.toString.length - i.toString.length) + i.toString
 
 }
